Add reusable hooks and refactor PlayerTable
