+++ src/components/table/PlayerTable.jsx
@@ -1,6 +1,6 @@
-import React, { useEffect, useState, useMemo, useCallback } from 'react';
-import { collection, getDocs } from 'firebase/firestore';
-import { db } from '@/firebaseConfig';
+import React, { useState, useMemo } from 'react';
+import usePlayerData from '@/hooks/usePlayerData';
+import useFilteredPlayers from '@/hooks/useFilteredPlayers';
 import PlayerRow from '@/components/table/PlayerRow';
 import FilterPanel from '@/components/filters/FilterPanel';
 import ActiveFiltersDisplay from '@/components/filters/ActiveFiltersDisplay';
@@ -8,56 +8,6 @@ import ViewControls from '@/components/filters/sections/ViewControls';
 import { Filter, SortAsc, Search } from 'lucide-react';
 import debounce from 'lodash.debounce';
 
-const positionAbbreviations = {
-  'Point Guard': 'PG',
-  'Shooting Guard': 'SG',
-  Guard: 'G',
-  'Small Forward': 'SF',
-  'Power Forward': 'PF',
-  Forward: 'F',
-  Center: 'C',
-  'Forward-Center': 'F/C',
-  'Guard-Forward': 'G/F',
-  'Forward-Guard': 'F',
-  'Center-Forward': 'C',
-};
-
-const expandPositionGroup = (position) => {
-  switch (position) {
-    case 'Guard':
-      return ['G'];
-    case 'Wing':
-      return ['G/F', 'F'];
-    case 'Forward':
-      return ['F', 'F/C'];
-    case 'Big':
-      return ['F/C', 'C'];
-    case 'Center':
-      return ['C'];
-    default:
-      return position ? [position] : [];
-  }
-};
-
-const shootingProfileRank = {
-  Elite: 6,
-  Plus: 5,
-  Capable: 4,
-  Willing: 3,
-  Hesitant: 2,
-  Non: 1,
-};
-
-const traitSort = [
-  'Defense',
-  'Energy',
-  'Feel',
-  'IQ',
-  'Passing',
-  'Playmaking',
-  'Rebounding',
-  'Shooting',
-];
 
 const getDefaultFilters = () => ({
   nameSearch: '',
@@ -119,15 +69,10 @@ const getDefaultFilters = () => ({
   badges: [],
 });
 
-const calculateHeight = (ht = '0-0') => {
-  const parts = ht.split('-');
-  return parseInt(parts[0]) * 12 + parseInt(parts[1]);
-};
 
 const PlayerTable = () => {
   const [filters, setFilters] = useState(getDefaultFilters());
-  const [players, setPlayers] = useState([]);
-  const [loading, setLoading] = useState(true);
+  const { players, loading } = usePlayerData();
   const [showFilters, setShowFilters] = useState(false);
   const [showSort, setShowSort] = useState(false);
 
@@ -145,309 +90,10 @@ const PlayerTable = () => {
       }, 200),
     []
   );
+  const filteredPlayers = useFilteredPlayers(players, filters);
 
-  const processPlayerData = useCallback((playerData) => {
-    const rawPosition = playerData.bio?.Position;
-    const formattedPosition =
-      positionAbbreviations[rawPosition] || rawPosition || '—';
 
-    const salaryMap = {};
-    (playerData.contract?.annual_salaries || []).forEach((s) => {
-      let raw = s.salary;
-      if (typeof raw === 'string') {
-        raw = raw.replace(/[\$,]/g, '');
-        if (raw.includes('M')) {
-          raw = raw.replace('M', '');
-          salaryMap[s.year] = parseFloat(raw);
-        } else {
-          salaryMap[s.year] = parseFloat(raw) / 1_000_000;
-        }
-      } else if (typeof raw === 'number') {
-        salaryMap[s.year] = raw / 1_000_000;
-      } else {
-        salaryMap[s.year] = null;
-      }
-      if (isNaN(salaryMap[s.year])) salaryMap[s.year] = null;
-    });
 
-    return {
-      ...playerData,
-      formattedPosition,
-      heightInInches: calculateHeight(playerData.bio?.HT),
-      weight: parseInt(playerData.bio?.WT || 0),
-      age: parseInt(playerData.bio?.AGE || 0),
-      headshotUrl: `/assets/headshots/${playerData.player_id}.png`,
-      offenseRole: playerData.roles?.offense1 || '—',
-      defenseRole: playerData.roles?.defense1 || '—',
-      shootingProfile: playerData.shootingProfile || '—',
-      subRoles: {
-        offense: playerData.subRoles?.offense || [],
-        defense: playerData.subRoles?.defense || [],
-      },
-      traits: playerData.traits || {},
-      badges: playerData.badges || [],
-      salaryByYear: salaryMap,
-      PPG: playerData.system?.stats?.PTS ?? null,
-      RPG: playerData.system?.stats?.TRB ?? null,
-      APG: playerData.system?.stats?.AST ?? null,
-      ...playerData.system?.stats,
-    };
-  }, []);
-
-  useEffect(() => {
-    const fetchPlayers = async () => {
-      try {
-        const querySnapshot = await getDocs(collection(db, 'players'));
-        const fetchedPlayers = querySnapshot.docs.map((doc) =>
-          processPlayerData(doc.data())
-        );
-        setPlayers(fetchedPlayers);
-      } catch (err) {
-        console.error('Error fetching players:', err);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchPlayers();
-  }, [processPlayerData]);
-
-  const filteredPlayers = useMemo(() => {
-    const selectedPositions = expandPositionGroup(filters.position);
-
-    return players
-      .filter((p) => {
-        // Name search filter
-        if (filters.nameSearch) {
-          const playerName = (p.display_name || p.name || '').toLowerCase();
-          const searchTerm = filters.nameSearch.toLowerCase();
-          if (!playerName.includes(searchTerm)) {
-            return false;
-          }
-        }
-
-        // Early exit for basic filters
-        if (
-          filters.team &&
-          (p.bio?.Team || '').toLowerCase() !== filters.team.toLowerCase()
-        ) {
-          return false;
-        }
-
-        if (
-          filters.position &&
-          !selectedPositions.includes(p.formattedPosition)
-        ) {
-          return false;
-        }
-
-        // Height filter
-        if (
-          p.heightInInches < filters.minHeight ||
-          (filters.maxHeight !== null && p.heightInInches > filters.maxHeight)
-        ) {
-          return false;
-        }
-
-        // Weight filter
-        if (
-          p.weight < filters.minWeight ||
-          (filters.maxWeight !== null && p.weight > filters.maxWeight)
-        ) {
-          return false;
-        }
-
-        // Age filter
-        if (
-          p.age < filters.minAge ||
-          (filters.maxAge !== null && p.age > filters.maxAge)
-        ) {
-          return false;
-        }
-
-        // Salary filter
-        if (
-          filters.minSalary !== undefined ||
-          filters.maxSalary !== undefined
-        ) {
-          const salary = p.salaryByYear?.[filters.salaryYear] ?? null;
-          if (typeof salary !== 'number') return false;
-          if (filters.minSalary !== undefined && salary < filters.minSalary)
-            return false;
-          if (filters.maxSalary !== undefined && salary > filters.maxSalary)
-            return false;
-        }
-
-        // Free agent filters
-        if (
-          filters.freeAgentYear &&
-          parseInt(p.free_agency_year || 0) !== parseInt(filters.freeAgentYear)
-        ) {
-          return false;
-        }
-
-        if (
-          filters.freeAgentType &&
-          p.free_agent_type !== filters.freeAgentType
-        ) {
-          return false;
-        }
-
-        // Role filters
-        if (
-          filters.offenseRole &&
-          ![p.roles?.offense1, p.roles?.offense2]
-            .filter(Boolean)
-            .some((role) =>
-              role.toLowerCase().includes(filters.offenseRole.toLowerCase())
-            )
-        ) {
-          return false;
-        }
-
-        if (
-          filters.defenseRole &&
-          ![p.roles?.defense1, p.roles?.defense2]
-            .filter(Boolean)
-            .some((role) =>
-              role.toLowerCase().includes(filters.defenseRole.toLowerCase())
-            )
-        ) {
-          return false;
-        }
-
-        // Shooting profile
-        if (
-          filters.shootingProfile &&
-          p.shootingProfile !== filters.shootingProfile
-        ) {
-          return false;
-        }
-
-        // Subroles filter
-        if (
-          filters.subRoles?.offense?.length ||
-          filters.subRoles?.defense?.length
-        ) {
-          if (
-            filters.subRoles.offense?.length &&
-            !filters.subRoles.offense.every((sub) =>
-              p.subRoles.offense.includes(sub)
-            )
-          ) {
-            return false;
-          }
-
-          if (
-            filters.subRoles.defense?.length &&
-            !filters.subRoles.defense.every((sub) =>
-              p.subRoles.defense.includes(sub)
-            )
-          ) {
-            return false;
-          }
-        }
-
-        // Stats filters
-        const passesStat = (key, min, max) => {
-          const val = parseFloat(p[key] ?? 0) * (key.includes('%') ? 100 : 1);
-          return val >= filters[`min_${min}`] && val <= filters[`max_${max}`];
-        };
-
-        if (
-          !passesStat('PTS', 'PPG', 'PPG') ||
-          !passesStat('TRB', 'RPG', 'RPG') ||
-          !passesStat('AST', 'APG', 'APG') ||
-          !passesStat('FG%', 'FGP', 'FGP') ||
-          !passesStat('3P%', 'TPP', 'TPP') ||
-          !passesStat('FT%', 'FTP', 'FTP') ||
-          !passesStat('eFG%', 'eFGP', 'eFGP') ||
-          !passesStat('MP', 'MIN', 'MIN') ||
-          !passesStat('G', 'G', 'G')
-        ) {
-          return false;
-        }
-
-        // Traits filters
-        const passesTrait = (trait) => {
-          const val = parseFloat(p.traits?.[trait] ?? 0);
-          return (
-            val >= filters[`min_${trait}`] && val <= filters[`max_${trait}`]
-          );
-        };
-
-        if (
-          !passesTrait('Defense') ||
-          !passesTrait('Energy') ||
-          !passesTrait('Feel') ||
-          !passesTrait('IQ') ||
-          !passesTrait('Passing') ||
-          !passesTrait('Playmaking') ||
-          !passesTrait('Rebounding') ||
-          !passesTrait('Shooting')
-        ) {
-          return false;
-        }
-
-        // Badges filter
-        if (
-          filters.badges?.length &&
-          !filters.badges.every((b) => p.badges.includes(b))
-        ) {
-          return false;
-        }
-
-        return true;
-      })
-      .sort((a, b) => {
-        const getValue = (player, field) => {
-          if (traitSort.includes(field)) return player.traits?.[field] ?? -1;
-          if (player.system?.stats?.hasOwnProperty(field))
-            return parseFloat(player.system.stats[field]) ?? -1;
-          switch (field) {
-            case 'name':
-              return (player.display_name || player.name || '').toLowerCase();
-            case 'height':
-              return player.heightInInches;
-            case 'weight':
-              return player.weight;
-            case 'age':
-              return player.age;
-            case 'salary':
-              return player.salaryByYear?.[filters.salaryYear] ?? -1;
-            case 'shootingProfile':
-              return shootingProfileRank[player.shootingProfile] ?? 0;
-            case 'yearsRemaining':
-              return parseInt(player.free_agency_year) - 2024 || -1;
-            case 'totalContract':
-              return Array.isArray(player.contract?.annual_salaries)
-                ? player.contract.annual_salaries.reduce((sum, s) => {
-                    const val =
-                      typeof s.salary === 'number'
-                        ? s.salary
-                        : parseFloat((s.salary || '').replace(/[^0-9.]/g, ''));
-                    return isNaN(val) ? sum : sum + val;
-                  }, 0)
-                : -1;
-            case 'overall':
-              return parseFloat(player.overall) ?? -1;
-            default:
-              return -1;
-          }
-        };
-
-        const valA = getValue(a, filters.sortBy);
-        const valB = getValue(b, filters.sortBy);
-
-        if (typeof valA === 'string' && typeof valB === 'string') {
-          return filters.sortAsc
-            ? valA.localeCompare(valB)
-            : valB.localeCompare(valA);
-        } else {
-          return filters.sortAsc ? valA - valB : valB - valA;
-        }
-      });
-  }, [players, filters]);
 
   const handleSearchChange = (e) => {
     debouncedSearchUpdate(e.target.value);
